Rename program to chack_staking
