+++ anchor/src/chack-staking-exports.ts
@@ -1,20 +1,20 @@
 // Here we export some useful types and functions for interacting with the Anchor program.
 import { Cluster, PublicKey } from '@solana/web3.js';
 import { Program } from '@coral-xyz/anchor'
-import type { Counter } from '../target/types/counter';
-import { IDL as CounterIDL } from '../target/types/counter';
+import type { ChackStaking } from '../target/types/chack_staking';
+import { IDL as ChackStakingIDL } from '../target/types/chack_staking';
 
 // Re-export the generated IDL and type
-export { Counter, CounterIDL };
-export type CounterProgram = Program<Counter>;
+export { ChackStaking, ChackStakingIDL };
+export type ChackStakingProgram = Program<ChackStaking>;
 
 // After updating your program ID (e.g. after running `anchor keys sync`) update the value below.
-export const COUNTER_PROGRAM_ID = new PublicKey(
+export const CHACK_STAKING_PROGRAM_ID = new PublicKey(
   'CounNZdmsQmWh7uVngV9FXW2dZ6zAgbJyYsvBpqbykg'
 );
 
 // This is a helper function to get the program ID for the Counter program depending on the cluster.
-export function getCounterProgramId(cluster: Cluster) {
+export function getChackStakingProgramId(cluster: Cluster) {
   switch (cluster) {
     case 'devnet':
     case 'testnet':
@@ -22,6 +22,6 @@ export function getCounterProgramId(cluster: Cluster) {
       // You only need to update this if you deploy your program on one of these clusters.
       return new PublicKey('CounNZdmsQmWh7uVngV9FXW2dZ6zAgbJyYsvBpqbykg');
     default:
-      return COUNTER_PROGRAM_ID;
+      return CHACK_STAKING_PROGRAM_ID;
   }
 }
