+++ anchor/src/index.ts
@@ -1,1 +1,1 @@
-export * from './counter-exports';
+export * from './chack-staking-exports';
